requirements.txt
faiss-cpu
numpy
flask
streamlit
langchain
sentence-transformers
transformers
requests
python-dotenv
